feat(support): show signups per day as an average on the tile
The four signup counters showed the raw total for each window, which is not
comparable across windows. They now show signups per day, averaged over the
last day, week, month and quarter, with one decimal.

SupportChartMini counters take a num so a fractional average renders with one
decimal while whole values keep showing no decimals.

+++ flutter/packages/growerp_models/lib/src/models/support_dashboard_model.dart
@@ -94,10 +94,11 @@ abstract class SupportSignupsStats with _$SupportSignupsStats {
   SupportSignupsStats._();
   factory SupportSignupsStats({
     @Default([]) List<SupportBarItem> bars,
-    @Default(0) int day,
-    @Default(0) int week,
-    @Default(0) int month,
-    @Default(0) int quarter,
+    // signups per day, averaged over the last day, week, month and quarter
+    @Default(0) double day,
+    @Default(0) double week,
+    @Default(0) double month,
+    @Default(0) double quarter,
   }) = _SupportSignupsStats;
 
   factory SupportSignupsStats.fromJson(Map<String, dynamic> json) =>
